third_party/WebKit/LayoutTests/fast/borders/border-color-visited.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<style>
.ui-btn-up-c
{
    border: 1px solid #d0d0d0;
    background: #ffffff;
    color: #444444;
}

.ui-link-inherit 
{
    text-decoration: none;
    white-space: nowrap;
}

.ui-btn-up-c:visited, .ui-btn-up-c a.ui-link-inherit
{
    color: #444444;
}
</style>
</head> 
<body>
<ul>
<iframe src="../history/resources/dummy.html" style="display:none"></iframe>
<li class="ui-btn-up-c">
  <span class="ui-btn-up-c">This span should have borders with #d0d0d0 color.</span>
</li>
<li class="ui-btn-up-c">
  <a href="../history/resources/dummy.html" class="ui-link-inherit">
    <span class="ui-btn-up-c">This span should also have borders with #d0d0d0 color.</span>
  </a>
</li>
</ul>
<div id="console"></div>
</body>
<script>
description('If this test passes, all borders have the same color: #d0d0d0.');
var spans = document.getElementsByTagName('span');
shouldBe('window.getComputedStyle(spans[0]).borderColor', '"rgb(208, 208, 208)"');
shouldBe('window.getComputedStyle(spans[1]).borderColor', '"rgb(208, 208, 208)"');
</script>
</html>
