third_party/WebKit/LayoutTests/fast/borders/border-color-visited-expected.txt

This span should have borders with #d0d0d0 color.
This span should also have borders with #d0d0d0 color.

If this test passes, all borders have the same color: #d0d0d0.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(spans[0]).borderColor is "rgb(208, 208, 208)"
PASS window.getComputedStyle(spans[1]).borderColor is "rgb(208, 208, 208)"
PASS successfullyParsed is true

TEST COMPLETE
